requirements.txt
pandas
sentence-transformers
faiss-cpu
networkx
langchain
langchain-openai
langchain-community
tiktoken
langchainhub 